requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asqlite==1.0.0
astroid==2.9.3
asttokens==2.0.5
async-timeout==3.0.1
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
black==21.12b0
bleach==4.1.0
blinker==1.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.0.3
cloud-init==21.4
colorama==0.4.3
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
devscripts===2.20.2ubuntu2
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
gpg===1.13.1-unknown
greenlet==0.4.15
httplib2==0.14.0
humanfriendly==10.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
ipykernel==6.8.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.18.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
MarkupSafe==1.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==4.2.0
msgpack==0.6.2
multidict==5.2.0
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
neotermcolor==2.0.10
nest-asyncio==1.5.4
netifaces==0.10.4
notebook==6.4.8
numpy==1.22.1
oauthlib==3.1.0
openai==0.13.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
pandas-stubs==1.2.0.45
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.6.0
pickleshare==0.7.5
pip-autoremove==0.10.0
platformdirs==2.4.1
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cord==2.0.0a0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.12.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyrsistent==0.15.5
pyserial==3.4
pyte==0.8.0
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
python-magic==0.4.16
pytz==2021.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
shtab==1.5.3
simplejson==3.16.0
six==1.14.0
sos==4.2
ssh-import-id==5.10
stack-data==0.1.4
systemd-python==234
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
thefuck==3.32
tldr==3.0.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
Twisted==18.9.0
typing-extensions==4.0.1
ubuntu-advantage-tools==27.5
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.3.2
unidiff==0.5.5
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wrapt==1.13.3
yarl==1.7.2
youtube-dl==2021.12.17
zipp==1.0.0
zope.interface==4.7.1
